% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{check_pal_n_RR}
\alias{check_pal_n_RR}
\title{Check palette size (help function)}
\usage{
check_pal_n_RR(n, max_n)
}
\arguments{
\item{n}{Number of discrete colors in the palette}

\item{max_n}{Maximum possible number of colors in the palette}
}
\description{
Check palette size (help function)
}
