% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-themes.R
\name{geom_bar_flipped_RR}
\alias{geom_bar_flipped_RR}
\title{Custom wrapper for flipped version of geom_bar}
\usage{
geom_bar_flipped_RR()
}
\description{
Custom wrapper for flipped version of geom_bar
}
\examples{
library(ggplot2)
mtcars |>
ggplot(aes(x = factor(cyl),
fill = factor(am))) +
theme_RR() +
geom_bar_flipped_RR() +
labs(title = "Number of Cars per Number of Cylinders",
x = "Number of Cylinders",
y = "Count") +
scale_fill_RR()

}
