% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_RR.R
\name{geom_bar_RR}
\alias{geom_bar_RR}
\title{Custom wrapper for geom_bar}
\usage{
geom_bar_RR(..., width = 0.6)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{width}{Bar width. By default, set to 90\% of the \code{\link[ggplot2:resolution]{resolution()}} of the
data.}
}
\description{
Custom wrapper for geom_bar
}
