% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific-themes.R
\name{geom_jitter_RR}
\alias{geom_jitter_RR}
\title{Custom wrapper for geom_jitter}
\usage{
geom_jitter_RR(..., size = 2.5, width = 0.4, height = 0.4)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{size}{Size of the dots.}

\item{width}{Amount of horizontal jitter.}

\item{height}{Amount of vertical jitter.}
}
\description{
Custom wrapper for geom_jitter
}
\examples{
library(ggplot2)
iris |>
ggplot(aes(x = Species,
y = Sepal.Length)) +
theme_RR() +
geom_jitter_RR(width = 0.1)

}
