% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{RR_pal}
\alias{RR_pal}
\title{RR color palette
Defines the official color palette of the Norwegian Office of the Auditor General.}
\usage{
RR_pal(fill = TRUE)
}
\arguments{
\item{fill}{Determines if the palette is used in a fill or color scale function.}
}
\description{
RR color palette
Defines the official color palette of the Norwegian Office of the Auditor General.
}
\examples{
library(scales)
show_col(RR_pal()(9))
}
