% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{RR_pal_alternative}
\alias{RR_pal_alternative}
\title{RR color palette (alternative)
Defines an alternative order of the official color palette of the Norwegian Office of the Auditor General.}
\usage{
RR_pal_alternative(fill = TRUE)
}
\arguments{
\item{fill}{Determines if the palette is used in a fill or color scale function.}
}
\description{
RR color palette (alternative)
Defines an alternative order of the official color palette of the Norwegian Office of the Auditor General.
}
\examples{
library(scales)
show_col(RR_pal_alternative()(10))
}
