% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{RR_trafikk_pal_green}
\alias{RR_trafikk_pal_green}
\title{RR traffic color palette (green)
Defines the official traffic color palette of the Norwegian Office of the Auditor General.}
\usage{
RR_trafikk_pal_green(fill = TRUE)
}
\arguments{
\item{fill}{Determines if the palette is used in a fill or color scale function.}
}
\description{
RR traffic color palette (green)
Defines the official traffic color palette of the Norwegian Office of the Auditor General.
}
\examples{
library(scales)
show_col(RR_trafikk_pal_green()(5))
}
