% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_RR.R
\name{scale_color_continuous_RR_blue}
\alias{scale_color_continuous_RR_blue}
\title{Continuous color scale (blue)}
\usage{
scale_color_continuous_RR_blue(...)
}
\arguments{
\item{...}{}
}
\description{
Continuous color scale (blue)
}
