% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_RR.R
\name{scale_color_continuous_RR_red}
\alias{scale_color_continuous_RR_red}
\title{Continuous color scale (red)}
\usage{
scale_color_continuous_RR_red(...)
}
\arguments{
\item{...}{}
}
\description{
Continuous color scale (red)
}
