% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_RR.R
\name{theme_RR}
\alias{theme_RR}
\title{Basic theme function to implement the official graph style of the Norwegian Office of the Auditor General.}
\usage{
theme_RR(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Basic theme function to implement the official graph style of the Norwegian Office of the Auditor General.
}
\examples{
library(ggplot2)
p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg,
colour = factor(gear)))

p + theme_RR()
}
