% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_RR.R
\name{geom_density_RR}
\alias{geom_density_RR}
\title{Custom wrapper for geom_density}
\usage{
geom_density_RR(..., alpha = 0.5)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{alpha}{Degree of transparency in the density plots.}
}
\description{
Custom wrapper for geom_density
}
