% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R, R/scale.R, R/specific-themes.R,
%   R/theme_RR.R
\name{imports}
\alias{imports}
\title{import ggplot2
import scales
import ggthemes
import sysfonts
import showtext}
\description{
import ggplot2
import scales
import ggthemes
import sysfonts
import showtext
}
